Move story to Editing Examples
