+++ packages/mantine-react-table/stories/features/EmptyRow.stories.tsx
@@ -103,56 +103,3 @@ export const EmptyRowExplanationPannel = () => {
 
   return <MantineReactTable table={table} />;
 };
-
-const sampleData: Person[] = [
-  {
-    firstName: 'Alice',
-    lastName: 'Smith',
-  },
-  {
-    firstName: 'Bob',
-    lastName: 'Johnson',
-  },
-  {
-    firstName: 'Charlie',
-    lastName: 'Williams',
-  },
-];
-
-export const AddingOrEditingInDetailPannel = () => {
-  //Now that empty row is an actual row, detail pannel is available, and can ne used as a form, maybe?
-  const [withData, setWithData] = useState(false);
-
-  const table = useMantineReactTable({
-    columns,
-    data: withData ? sampleData: data,
-    renderDetailPanel: ({ table, row, internalEditComponents }) => (
-      <Center>
-        <form onSubmit={(e) => e.preventDefault()}>
-          <Group gap="md" pb={24} pt={16}>
-            {internalEditComponents}
-          </Group>
-        </form>
-        <Flex justify="flex-end">
-          <MRT_EditActionButtons row={row} table={table} variant="text" />
-        </Flex>
-      </Center>
-    ),
-    renderEmptyRowsFallback: () => (
-      <Center>
-        <Text>This table is empty, click on the chevron to add a record</Text>
-      </Center>
-    ),
-  });
-
-  return (
-    <Stack>
-      <Switch
-        checked={withData}
-        label="Show data"
-        onChange={(e) => setWithData(e.currentTarget.checked)}
-      />
-      <MantineReactTable table={table} />
-    </Stack>
-  );
-};
